requirements.txt
altgraph==0.17.4
asgiref==3.7.2
beautifulsoup4==4.12.2
build==1.0.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap5==2023.10
cryptography==41.0.5
dj-database-url==2.1.0
Django==4.2.6
django-bootstrap-v5==1.0.11
django-crispy-forms==2.1
django-cors-headers==4.3.0
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-filter==23.3
django-on-heroku==1.1.2
djangorestframework==3.14.0
docutils==0.20.1
googleapis-common-protos==1.61.0
grpcio==1.59.2
grpcio-status==1.59.2
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
jaraco.classes==3.3.0
jeepney==0.8.0
jws==0.1.3
keyring==24.2.0
Markdown==3.5.1
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==10.1.0
nh3==0.2.14
packaging==23.2
Pillow==10.1.0
pip-autoremove==0.10.0
pip-review==1.3.0
pkginfo==1.9.6
protobuf==4.25.0
psycopg==3.1.12
psycopg2-binary==2.9.9
pycparser==2.21
Pygments==2.16.1
pyinstaller==6.1.0
pyinstaller-hooks-contrib==2023.10
pyproject_hooks==1.0.0
pytz==2023.3.post1
readme-renderer==42.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.6.0
SecretStorage==3.3.3
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tensorflow-intel==0.0.1
tomli==2.0.1
twine==4.0.2
typing_extensions==4.8.0
tzdata==2023.3
uplink==0.9.7
uritemplate==4.1.1
urllib3==2.0.7
whitenoise==6.6.0
zipp==3.17.0
